[feature] Query string parameters

+++ Notes/Views/NotePage.xaml.cs
@@ -1,5 +1,6 @@
 namespace Notes.Views;
 
+[QueryProperty(nameof(ItemId), nameof(ItemId))]
 public partial class NotePage : ContentPage
 {
 	string _fileName = Path.Combine(FileSystem.AppDataDirectory, "notes.txt");
@@ -13,19 +14,32 @@ public partial class NotePage : ContentPage
 		LoadNote(Path.Combine(appDataPath, randomFileName));
 	}
 
-	private void SaveButton_Clicked(object sender, EventArgs e)
+	public string ItemId
 	{
-		// Save the file.
-		File.WriteAllText(_fileName, TextEditor.Text);
+		set
+		{
+			LoadNote(value);
+		}
 	}
 
-	private void DeleteButton_Clicked(object sender, EventArgs e)
+	private async void SaveButton_Clicked(object sender, EventArgs e)
 	{
-		// Delete the file.
-		if (File.Exists(_fileName))
-			File.Delete(_fileName);
+		if (BindingContext is Models.Note note)
+			File.WriteAllText(_fileName, TextEditor.Text);
 
-		TextEditor.Text = string.Empty;
+		await Shell.Current.GoToSync("..");
+	}
+
+	private async void DeleteButton_Clicked(object sender, EventArgs e)
+	{
+		if (BindingContext is Models.Note note)
+		{
+			// Delete the file.
+			if (File.Exists(_fileName))
+				File.Delete(_fileName);
+		}
+
+		await Shell.Current.GoToSync("..");
 	}
 
 	private void LoadNote(string fileName)
